Moved main loop stuff to it's own file
